+++ docs/changelog.rst
@@ -7,3 +7,6 @@ v0.1 (not yet released)
 * Initial network interface
 
 * Network implementations: `aig_network`, `klut_network`
+
+* Algorithms: `cut_enumeration`
+  `#2 <https://github.com/lsils/mockturtle/pull/2>`_